api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
redis>=5.0.0
minio>=7.2.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic-settings>=2.1.0
prometheus-client>=0.19.0
aiosmtplib>=3.0.0
stripe>=8.0.0
